Eliminate warning by casting size to ssize_t in check macro.

+++ hll.h
@@ -1103,7 +1103,7 @@ public:
     }
     void write(int fileno) const {
         uint32_t bf[]{is_calculated_, clamp_, estim_, jestim_, nthreads_};
-#define CHWR(fn, obj, sz) if(__builtin_expect(::write(fn, (obj), (sz)) != (sz), 0)) throw std::runtime_error(std::string("Failed to write to disk in ") + __PRETTY_FUNCTION__)
+#define CHWR(fn, obj, sz) if(__builtin_expect(::write(fn, (obj), (sz)) != ssize_t(sz), 0)) throw std::runtime_error(std::string("Failed to write to disk in ") + __PRETTY_FUNCTION__)
         CHWR(fileno, bf, sizeof(bf));
         CHWR(fileno, &np_, sizeof(np_));
         CHWR(fileno, &value_, sizeof(value_));
@@ -1112,7 +1112,7 @@ public:
     }
     void read(int fileno) {
         uint32_t bf[5];
-#define CHRE(fn, obj, sz) if(__builtin_expect(::read(fn, (obj), (sz)) != (sz), 0)) throw std::runtime_error(std::string("Failed to read from fd in ") + __PRETTY_FUNCTION__)
+#define CHRE(fn, obj, sz) if(__builtin_expect(::read(fn, (obj), (sz)) != ssize_t(sz), 0)) throw std::runtime_error(std::string("Failed to read from fd in ") + __PRETTY_FUNCTION__)
         CHRE(fileno, bf, sizeof(bf));
         is_calculated_ = bf[0];
         clamp_         = bf[1];
